+++ src/web/presentation/recruteur/serializers.py
@@ -10,6 +10,7 @@ from presentation.commons.serializers import LocalisationSerializer, OrganismeSe
 
 
 class OrganismeDetailSerializer(serializers.Serializer):
+    organisme_uuid = serializers.UUIDField()
     nom = serializers.CharField()
     siret = serializers.CharField(min_length=14, max_length=14)
     gestionnaire = serializers.CharField(allow_null=True)
